server/dbadmin/requirements.txt
flask==1.1.1
flask-sqlalchemy
flask-migrate
flask-cors
requests
SQLAlchemy-ImageAttach
prometheus-flask-exporter
flask_prometheus_metrics